+++ level_editor/src/org/vertexarmy/dsr/leveleditor/editors/sprite/SpriteEditor.java
@@ -3,6 +3,7 @@ package org.vertexarmy.dsr.leveleditor.editors.sprite;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Vector2;
 import lombok.Getter;
+import lombok.Setter;
 import org.vertexarmy.dsr.core.Root;
 import org.vertexarmy.dsr.core.assets.TextureRepository;
 import org.vertexarmy.dsr.core.component.ComponentType;
@@ -30,10 +31,11 @@ public class SpriteEditor {
     @Getter
     private GenericEditor editDialog;
 
-    private Root root;
+    @Getter
+    @Setter
+    private SpriteEditorListener listener;
 
     public SpriteEditor(Root root) {
-        this.root = root;
         node.addComponent(ComponentType.INPUT, new SpriteEditorInputComponent(this));
         node.addComponent(ComponentType.RENDER, new SpriteEditorRenderComponent(this));
 
@@ -65,11 +67,24 @@ public class SpriteEditor {
         return this.levelSprite != null;
     }
 
-    public Vector2 getSpriteBottomLeftCorner() {
+    void deleteSprite() {
+        if (isBoundToSprite()) {
+            notifyDeleteSpriteRequested();
+            unbindFromSprite();
+        }
+    }
+
+    private void notifyDeleteSpriteRequested() {
+        if (listener != null) {
+            listener.deleteSpriteRequested();
+        }
+    }
+
+    Vector2 getSpriteBottomLeftCorner() {
         return levelSprite.getPosition();
     }
 
-    public Vector2 getSpriteTopRightCorner() {
+    Vector2 getSpriteTopRightCorner() {
         TextureRegion region = TextureRepository.instance().getTexture(levelSprite.getTextureName());
         float width = region.getRegionWidth() * levelSprite.getScale().x;
         float height = region.getRegionHeight() * levelSprite.getScale().y;