+++ level_editor/src/org/vertexarmy/dsr/leveleditor/editors/sprite/SpriteEditorListener.java
@@ -1,0 +1,9 @@
+package org.vertexarmy.dsr.leveleditor.editors.sprite;
+
+/**
+ * created by Alex
+ * on 08-Apr-2015.
+ */
+public interface SpriteEditorListener {
+    void deleteSpriteRequested();
+}