+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Support for editing texture overlays, and drawing terrain patches with texture overlays
 - Spinner UI components
 - Support for removing terrain patches
+- Support for removing terrain sprites
 
 ## [1.2] 2015-04-04
 ### Added